new engine_data and error ffi modules
